requirements.txt
requests==2.28.1
Flask>=2.0
gradio-client
Werkzeug>=2.3.3
python-dotenv==1.0.0
cloudinary==1.32.0